hw/virtio/vhost-backend: Support vhost-iommu
When using vhost-iommu, virtio-iommu needs to register its endpoints
with an ioctl. Add that support to the kernel backend.

This is work in progress because at the moment I don't know how to
cleanly get from the endpoint to the vhost device that contains the fd
to register. At the moment we match them using the memory region, but
it's not generic.

Signed-off-by: Jean-Philippe Brucker <jean-philippe@linaro.org>